requirements.txt

streamlit>=1.36
lxml>=5.2
pydantic>=2.6
rapidfuzz>=3.9
pypdf>=4.2
python-docx>=1.1
google-generativeai>=0.7
tqdm>=4.66
